Add files via upload
German-Lang-Trainer2.py is an improved version of the original script. One needs to add the API key to work. The GUI shows much better on Windows operating systems. In addition to scraping, you can now press 'TRANSLATE' to select a file for translation in English by the AI (gemini-2.0-flash)